pyproject.toml
[project]
name = "nutritional-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.99.1",
    "openai-agents==0.2.4",
    "tavily-python>=0.7.11",
]
